Add basic doctesting
Add doctest as a dependency, and a basic example using 'divide'
in Pattern.hs.

Rename test/Spec.hs to test/Doctest.hs in order to reflect
that only doctests are implemented.

Show instances are needed for a handful of types, including
Position, Velocity, and Path. With more tests, Eq should
become necessary as well.

Doctests will not suffice, and hspec should be added at some point.