Add a scaleTime function
This allows to advance a pattern by a step in time.
This is useful first of all to avoid spawning a path
inside an enemy.

Furthermore, it's more composable to provide functions
for specifying the shape, and then this function for scaling
the shape outwards.

This kind of transformation is crucial in having complex patterns.

Tests were added in the documentation and the spec. Some property tests
would be useful to add to this.