Add a HUD
Add a HUD to display player health matching
the current color of the player.

The next thing to add would be actual damage logic for the player.

This required modifying the drawing routines to draw the hud.
There's potential for code reuse between the entity drawing and the hud
drawing.