Make enemies not take damage from own bullets
This seperates bullets into 2 categories, Player,
and Not Player. A side effect of this change is that
enemy bullets pass through other enemies, which can be changed
in the future, but isn't an undesirable property a priori.

This branches the game logic between Hud vs NoHud in preparation
for enemy bullets damaging the player.